fix: Fixing Bug like and create get user by id

+++ src/controller/Like.ts
@@ -3,31 +3,31 @@ import { Responsedata, responseDataFunction } from "../Scema/Response";
 import { Like } from "../type/like";
 import { prisma } from "../libs/prisma";
 
-export const ToggleLike = async (req: Request, res: Response) => {
-  try {
-    const { userId } = req.params;
-    const data = req.body as Like;
-    const response = await GetLikeByPostId(Number(userId));
-    let responseLike;
-    if (response.status !== 200) {
-      responseLike = await LikePost(data);
-    } else {
-      responseLike = await UnlikePost(response.data.id);
-    }
+// export const ToggleLike = async (req: Request, res: Response) => {
+//   try {
+//     const data = req.body as Like;
+//     const response = await GetLikeByPostId(data.userId, data.postId);
+//     let responseLike;
+//     if (response.status !== 200) {
+//       responseLike = await LikePost(data);
+//     } else {
+//       responseLike = await UnlikePost(response.data.id);
+//     }
 
-    return Responsedata(
-      { data: responseLike.data, message: responseLike.message, status: 200 },
-      res
-    );
-  } catch (error: any) {
-    return Responsedata({ data: {}, message: error.message, status: 500 }, res);
-  }
-};
-export const LikePost = async (data: Like) => {
+//     return Responsedata(
+//       { data: responseLike.data, message: responseLike.message, status: 200 },
+//       res
+//     );
+//   } catch (error: any) {
+//     return Responsedata({ data: {}, message: error.message, status: 500 }, res);
+//   }
+// };
+export const LikePost = async (req: Request, res: Response) => {
   try {
-    const dataLike = data;
+    const data = req.body as Like;
+
     const response = await prisma.like.create({
-      data: dataLike as Like,
+      data: data as Like,
     });
     return responseDataFunction({
       data: response,
@@ -38,11 +38,19 @@ export const LikePost = async (data: Like) => {
     return responseDataFunction({ data: {}, message: "error", status: 500 });
   }
 };
-export const UnlikePost = async (id: string) => {
+export const UnlikePost = async (req: Request, res: Response) => {
   try {
-    const response = await prisma.like.delete({
+    const { id } = req.params;
+    if (!id) {
+      return responseDataFunction({
+        data: {},
+        message: "Post ID is required",
+        status: 400,
+      });
+    }
+    await prisma.like.delete({
       where: {
-        id,
+        id: id as string,
       },
     });
     return responseDataFunction({
@@ -54,11 +62,12 @@ export const UnlikePost = async (id: string) => {
     return responseDataFunction({ data: {}, message: "error", status: 500 });
   }
 };
-export const GetLikeByPostId = async (id: number) => {
+export const GetLikeByPostId = async (id: number, postId: string) => {
   try {
     const response = await prisma.like.findFirst({
       where: {
         userId: id,
+        postId: postId,
       },
     });
     if (!response) {
